Update dot to synchronize with rustc's graphviz
There have been a lot of changes in the upstream graphviz crate,
which haven't made it down here. The two significant changes are
that the trait node and edges have been converted into associated
types, and that there's been a IntoCow trait added to simplify
some code. The rest of the changes were mainly synchronizing on
style.

This also bumps the version to 0.2.0 because it's a breaking change.